# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyMate Personal Finance API
  version: 0.1.0
  description: |
    MVP REST API contract for MoneyMate Personal Finance.

    All API responses use the same envelope shape:
    `{ "success": true, "data": {}, "meta": {}, "request_id": "req_xxx" }`.

    Error responses use:
    `{ "success": false, "error": { "code": "...", "message": "...", "details": [] }, "request_id": "req_xxx" }`.

    Authentication uses short-lived JWT access tokens in the `Authorization: Bearer <token>` header.
    Login also sets an HTTP-only `moneymate_refresh_token` cookie scoped to `/api/v1/auth`.
    Refresh token values are not returned in JSON responses.

    Financial safety: all market prices in the MVP are manual/mock. The API does not claim real-time prices and does not provide buy/sell recommendations.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/v1/dashboard/overview:
    get:
      tags: [Dashboard]
      summary: Dashboard overview totals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Net worth, portfolio, cash, and P/L summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewEnvelope"
              examples:
                seeded:
                  value:
                    success: true
                    data:
                      base_currency: IDR
                      total_net_worth: 13648703.72
                      total_portfolio_value: 13492813.72
                      total_cash: 155890
                      total_cost: 14382050.41
                      profit_loss_value: -889236.69
                      profit_loss_percent: -0.0618
                      best_performer:
                        ticker: SPY
                        name: S&P 500 ETF
                        profit_loss_percent: 0.0135
                      worst_performer:
                        ticker: BBRI
                        name: Bank Rakyat Indonesia
                        profit_loss_percent: -0.1425
                      last_updated_at: "2026-06-30T15:01:57+07:00"
                      price_disclaimer: Data bukan real-time. Harga berasal dari input manual/mock.
                      financial_disclaimer: Informasi ini bersifat pencatatan dan analisis, bukan nasihat investasi.
                    request_id: req_123
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/dashboard/asset-allocation:
    get:
      tags: [Dashboard]
      summary: Asset allocation by category
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Asset allocation rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetAllocationListEnvelope"
              examples:
                seeded:
                  value:
                    success: true
                    data:
                      - asset: Reksadana
                        value: 5722589.05
                        percent: 0.4193
                      - asset: Saham
                        value: 4641000
                        percent: 0.34
                      - asset: Emas
                        value: 1626124.8
                        percent: 0.1191
                      - asset: ETF
                        value: 1503099.87
                        percent: 0.1101
                      - asset: Cash
                        value: 155890
                        percent: 0.0114
                    request_id: req_123
  /api/v1/dashboard/performance:
    get:
      tags: [Dashboard]
      summary: Portfolio performance rows
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Performance rows with disclaimer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceEnvelope"
  /api/v1/dashboard/alerts:
    get:
      tags: [Dashboard]
      summary: Dashboard data quality and risk alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Alert rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token from `/api/v1/auth/login` or `/api/v1/auth/refresh`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: moneymate_refresh_token
      description: HTTP-only refresh cookie set by `/api/v1/auth/login`, scoped to `/api/v1/auth`.
...
  responses:
...
    UnauthorizedError:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unauthorized:
              value:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: Access token tidak valid
                request_id: req_123
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, error, request_id]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorBody"
        request_id:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INTERNAL_ERROR, VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT]
        message:
          type: string
        details:
          type: array
          items:
            type: string
...
    SuccessEnvelope:
      type: object
      required: [success, request_id]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Endpoint-specific response data.
        meta:
          description: Optional endpoint metadata.
        request_id:
          type: string
...
    Performer:
      type: object
      properties:
        ticker:
          type:
            - string
            - "null"
        name:
          type: string
        profit_loss_percent:
          type: number
    DashboardOverview:
      type: object
      required: [base_currency, total_net_worth, total_portfolio_value, total_cash, total_cost, profit_loss_value, profit_loss_percent, price_disclaimer, financial_disclaimer]
      properties:
        base_currency:
          type: string
        total_net_worth:
          type: number
        total_portfolio_value:
          type: number
        total_cash:
          type: number
        total_cost:
          type: number
        profit_loss_value:
          type: number
        profit_loss_percent:
          type: number
        best_performer:
          oneOf:
            - $ref: "#/components/schemas/Performer"
            - type: "null"
        worst_performer:
          oneOf:
            - $ref: "#/components/schemas/Performer"
            - type: "null"
        last_updated_at:
          type:
            - string
            - "null"
          format: date-time
        price_disclaimer:
          type: string
        financial_disclaimer:
          type: string
    DashboardOverviewEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/DashboardOverview"
    AssetAllocation:
      type: object
      required: [asset, value, percent]
      properties:
        asset:
          type: string
        value:
          type: number
        percent:
          type: number
    AssetAllocationListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/AssetAllocation"
    PerformanceEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, disclaimer]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      ticker:
                        type:
                          - string
                          - "null"
                      name:
                        type: string
                      total_cost:
                        type: number
                      current_value:
                        type: number
                      profit_loss_value:
                        type: number
                      profit_loss_percent:
                        type: number
                disclaimer:
                  type: string
    Alert:
      type: object
      required: [code, title, message, severity]
      properties:
        code:
          type: string
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error]
    AlertListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
...